fix (be:/sysinfo/types): fix missmatch type refresh rate from display_fastfetch

+++ src-tauri/src/sysinfo/types/display_fastfetch.rs
@@ -43,7 +43,7 @@ pub struct ManufactureDate {
 pub struct Output {
     pub width: i64,
     pub height: i64,
-    pub refresh_rate: i64,
+    pub refresh_rate: f64,
     pub drr_status: Option<serde_json::Value>,
 }
